Replace crypto lib with ece
Refactor directory structure
