--- packages/web/src/lib/crypto/ece/constants.ts
@@ -1,9 +1,0 @@
-export const KEY_LENGTH = 16
-export const NONCE_LENGTH = 12
-export const TAG_LENGTH = 16
-export const RECORD_SIZE = 1024 * 64
-
-export enum Mode {
-	Encrypt = 0,
-	Decrypt = 1,
-}
