--- packages/web/src/lib/crypto/ece/index.ts
@@ -1,5 +1,0 @@
-export * from "./constants"
-export * from "./generate"
-export * from "./keysmith"
-export * from "./main"
-export * from "./utils"
