--- packages/web/src/lib/crypto/index.ts
@@ -1,6 +1,0 @@
-import * as ece from "./ece"
-import type { Crypto } from "./type"
-
-export * from "./utils"
-export type { Crypto }
-export { ece }
