--- packages/web/src/lib/crypto/type.ts
@@ -1,4 +1,0 @@
-/** This module enables easier type-safe dynamic imports */
-import type * as crypto from "./index"
-
-export type Crypto = typeof crypto
