+++ packages/web/src/lib/ece/type.ts
@@ -1,0 +1,4 @@
+/** This module enables easier type-safe dynamic imports */
+import type * as ece from "./index"
+
+export type Ece = typeof ece
